requirements.txt
flask==3.0.2
flask-cors==4.0.0
chardet==5.2.0
python-magic-bin==0.4.14 